cmd import: remove the dedicated pkg/import and keep the code in package main
Terravalet has multiple commands, all except "import" in the main package.
This asymmetry was making the overall code not consistent and thus harder to comprehend.

It might be that one day we realize that it makes sense to have packages, but we are not at that level of complexity yet.

PCI-2092