terravalet: command-line: add back printing the version
I am a bit torn on this one. I don't like using https://github.com/alexflint/go-arg#version-strings because it prints the version on each `-h` invocation, which is, I think, confusing.

I could use the approach we use in fly_helper, but I want to keep it simple and just use MustParse() as entry point.

Thus, I think that what is proposed in this commit is the less worse of the various approaches.

As a bonus, we show how to embed a struct (see `type UpDown` in main.go) and how to use "testable examples" (see main_test.go).