build: ensure that all release tags are annotated tags
Git supports two types of tags: lightweight and annotated.
We assume that all tags beginning with "v" are release tags (eg: v1.2.3) and we enforce that release tags are annotated.
See "git help tag" for more background information.

NOTE: the README contains instructions on how to make a release and the Taskfile already contains a `release` target that will issue the correct type of tag.

Part of PCI-2275