release: actually push the git tag
This can be confusing. At the beginning I thought that this was the problem:
we were locally making an annotated tag but then the Github release process
was making a lightweight tag behind our backs (?). This might or might not be
the case (needs further investigation). What is sure is that I did push the
v0.7.0 tag and the CI is still failing.